Remove empty directories in the installation dir after update
If a directory was removed from one version of the app
to the next, all of the files inside the directory
would be removed but a hierarchy of empty directories would
be left behind.

On OS X this caused code signing checks to fail after updating
from a release which stored code signing-related files in
'Contents/_CodeSignature' for an embedded framework to
'Versions/4/_CodeSignature', since the old _CodeSignature dir
was never removed.

MD-20523
test: auto: TestFileUtils