Set deployment target to Mac OS X 10.6
The way we previously set the deployment target is no longer working, as recent
versions of CMake actually set the deployment target with the SDK. At least in
the combination CMake 3.4 + XCode 7.1 (SDK 10.11), this fails.

This patch changes it in a way that is supposed to work since ages.

MD-21785