Directly use bzip2's source code when compiling
When compiling minizip on Windows, we used to link against static
libraries. These have been updated for Visual Studio 2015.
Unfortunately the code is confused about whether to use static
or dynamic libraries. This causes the updater.exe to have a
dependency on bz2.dll, which isn't present on user PCs.

By adding the few source files directly to the list of source files
of minizip, we can avoid the requirement to use prebuilt binaries
in the first place. Hence we're no longer depending on the actual
compiler used for bzip2/minizip.

MD-22726